fix: move endpoint de upload de arquivos de queimadas para escopo dos scripts
