+++ server/api/controllers/queimadas/router.ts
@@ -1,9 +1,5 @@
 import { Router } from 'express';
 import controller from './controller';
-import multer from 'multer';
-import os from 'os';
-
-const upload = multer({ dest: os.tmpdir() });
 
 import { middleware as cacheMiddleware } from 'apicache';
 
@@ -12,5 +8,4 @@ export default Router()
   .get('/:source/municipios', cacheMiddleware('1 week'), controller.find)
   .get('/:source/municipios/:municipio', cacheMiddleware('1 week'), controller.get)
   .get('/:source/estados', cacheMiddleware('1 week'), controller.find)
-  .get('/:source/estados/:estado', cacheMiddleware('1 week'), controller.get)
-  .post('/upload', upload.single('zipfile'), controller.upload);
+  .get('/:source/estados/:estado', cacheMiddleware('1 week'), controller.get);
